fix: make all veg codes usable unless it's for registration
