+++ frontend/src/views/Seedlot/ContextContainerClassA/index.tsx
@@ -114,7 +114,8 @@ const ContextContainerClassA = ({ children }: props) => {
 
   const vegCodeQuery = useQuery({
     queryKey: ['vegetation-codes'],
-    queryFn: () => getVegCodes(true),
+    queryFn: () => getVegCodes(),
+    select: (data) => getMultiOptList(data, true, true),
     staleTime: THREE_HOURS,
     cacheTime: THREE_HALF_HOURS
   });
